Feature: Allow a HumanError instance to 'look up' an error by name
This is so that we can configure each instance of an error more easily.

We currently are not rescuing the `NameError` that will be raised if the error
name provided does not exist:

For example:

    HumanError.new.fetch('foo')

would throw a `NameError`.  We should probably handle that in a more graceful
way, even if it means wrapping it in something like
`HumanError::ErrorLookupError`.

--------------------------------------------------------------------------------
Change-Id:     Ie071ef0b23e6c74dddb1a0108d5cb03c92ff3522
Signed-off-by: Jeff Felchner <accounts+git@thekompanee.com>