Feature: Allow local configuration to be passed to looked up errors
This uses `Configuration#to_h` to generate a set of initialization values that
are passed into each error that is looked up via `HumanError#fetch`.

This allows each instance of `HumanError` to create errors with different
information, even if they are of the same type.

Example:

  error = HumanError.new do |config|
            config.api_version = 1
          end

  error.fetch('RequestError').api_version # => 1

but if we created a new `HumanError` instance with an `api_version` of `2` then
calling `fetch('RequestError').api_version` on that instance would yield `2`.

--------------------------------------------------------------------------------
Change-Id:     I0c26b3f59af2588dbd52384bdadd8bce0ee5180d
Signed-off-by: Jeff Felchner <accounts+git@thekompanee.com>