Feature: Allow HumanError::Errors to have messages explicitly set
Previously the message that was displayed was divined from a set of variables
which was passed into the initializer, however there are times when we want to
be able to explicitly say what we want the message for the error to be.

Now we can pass in a `message` option when the error is created and it will
always use that exact text.

Example:

    error = MyHumanError.new(message: 'My custom error message')
    error.message # => 'My custom error message'

--------------------------------------------------------------------------------
Change-Id:     I10b1491f2e4aa1fbb332b5dcf419bb37588b24eb
Signed-off-by: Jeff Felchner and Drew Montgomery <pair+hhd418-jfelchner@example.com>