Feature: Add AssociationError
This error will be used when we have an issue associating one resource with
another.  For example, if we have a `Plan` and a `Subscription`, and there is no
plan with an ID of 123, if we attempt to associate a subscription with plan 123,
then it could potentially raise this type of error.

The first person to complain that the developer details keys isn't underscored
will get 30 lashes with a Singaporian cane (PS: Make sure Blake brings back
a Singaporian cane.) (PSS: Make sure Blake comes back period.)

--------------------------------------------------------------------------------
Change-Id:     I0de18b1569ded3debd700a8ae825e4695bf82da6
Signed-off-by: Jeff Felchner <accounts+git@thekompanee.com>