Bugfix: Rework the 'message' and 'to_s' methods on Error
When we were using a new class and mixing in `HumanError`, calling `to_s` was
blowing up (mainly when we were running specs).  We needed to make sure that
even when a new error is generated, that it can show its message, even if one
has not been set (or if `developer_message` has not been overridden).

--------------------------------------------------------------------------------
Change-Id:     Ia3671212ccfb71fc8fa100ac13c43b7737430072
Signed-off-by: Jeff Felchner <accounts+git@thekompanee.com>