Feature: Add HumanError::Persistable
We realized that there was a lot of common logic surrounding errors and wanted
to create an easier way for our custom `HumanError:Errors`.

We are now overriding the `save!` method inside of active record to catch the
following errors:

`ActiveRecord::InvalidForeignKey` errors will  raise a custom human error
`AssociationError`. Note for this error we also had to refactor the engine
itself to create foreign key relationships on the server level

`ActiveRecord::RecordInvalid` && `ActiveRecord::RecordNotSaved` errors will
raise a custom human error `ResourcePersistanceError`

We are also overriding the `find(*ids)` to catch the following errors:

`ActiveRecord:RecordNotFound` errors will raise a custom human error
`ResourceNotFoundError`

This allows us to include the `HumanError::Persistable` module in our different
engines and we can now delete a bunch of code!!

--------------------------------------------------------------------------------
Change-Id:     Ie08b95a8511c02e0e79dda95e95653882903728e
Signed-off-by: Jeff Felchner <accounts+git@thekompanee.com>