Refactor: Change the resolution of human error Persistable error names
Previously we were defining the `human_error_resource_name` as both
a module-level and instance-level method, but once we only wanted to include the
`ClassMethods` module without the enclosing `Persistable` module, we no longer
had access to it.

By defining it specifically as a module method, we can access it statically both
from the class, module or instance.

--------------------------------------------------------------------------------
Change-Id:     I0d6216f30990598ca7abbf40a9cb12d0450793ea
Signed-off-by: Jeff Felchner <accounts+git@thekompanee.com>