Removed: HumanError::Persistable
This module did a bunch of monkeypatching to `ActiveRecord` objects.  This also
meant that if you used the `ActiveRecord` objects outside the context of an API
controller, you would still get the `HumanError` errors which is probably not
what you want.

The way that it works now is that if you are executing via an API controller
with `RescuableResource` mixed in, it will transform errors into `HumanError`
errors.  Otherwise (say if you're just using an `ActiveRecord` object from
console or from Sidekiq), you will get the typical `ActiveRecord` errors.

The final benefit that I can think of for doing this is that adding the handling
of new errors into the mix is going to be substantially easier.

--------------------------------------------------------------------------------
Change-Id:     Iaf4b7e08bcd9a9c474735cc2694601293ddd9e1e
Signed-off-by: Jeff Felchner <accounts+git@thekompanee.com>